Refactor query rewriting to borrow unchanged queries
Return None for unchanged queries and reduce recursive Boolean rewrite stack usage. Align BooleanQuery and ConstantScoreQuery rewrites with Lucene 10.1, preserve signed minimum-should-match comparisons and Java integer wrapping, and restore deeply nested rewrite tests.

+++ src/core/search/wand_scorer.rs
@@ -247,7 +247,7 @@ where
   ) -> Result<DocIdSetIteratorImpl<S>> {
     let num_scorers = scorers.len();
 
-    if min_should_match as usize >= num_scorers {
+    if i64::from(min_should_match) >= num_scorers as i64 {
       return Err(LuceneError::illegal_argument(
         "minShouldMatch should be < the number of scorers",
       ));
